refactor(route53): split service.rs by resource family (4443 -> 1644 lines)
Audit 2026-05-19 P5. 65-method service.rs split into 7 family
submodules under service/:

- traffic_policies.rs   789
- mod.rs               1644  (dispatch + helpers + utility methods)
- dnssec.rs             445
- health_checks.rs      414
- hosted_zones.rs       411
- records.rs            370
- cidr.rs               312
- query_logging.rs      167

Same approach as PR #1479 / #1480: pub(super) family methods, #[path]
for helpers, mod tests stays in mod.rs. 38 unit tests pass.
